Move `mime.ts`
